requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
filetype==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
multidict==6.0.5
pillow==10.4.0
pyotp==2.9.0
setuptools==69.5.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
twikit==2.0.2
wheel==0.43.0
yarl==1.9.4
